avoid key package name
